Revert "Merge remote-tracking branch 'upstream/master'"
This reverts commit 70a4bb7007032439fdd0a0db839b61f386dae806, reversing
changes made to 2fea677e985bbb6fb41c66966a88d3d9c0937aad.

+++ PKHeX.Core/PersonalInfo/Interfaces/IPersonalEncounter.cs
@@ -18,7 +18,7 @@ public interface IPersonalEncounter
     /// <summary>
     /// Catch Rate
     /// </summary>
-    byte CatchRate { get; set; }
+    int CatchRate { get; set; }
 
     /// <summary>
     /// Initial Friendship when captured or received.
